% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_fun_cpp}
\alias{G_fun_cpp}
\title{G_fun_cpp}
\usage{
G_fun_cpp(x, y, W)
}
\arguments{
\item{x}{is a vector of length \eqn{2m}}

\item{y}{is a vector of length \eqn{2m}}

\item{W}{is a frequency matrix \eqn{m \times s}}
}
\value{
a vector of length \eqn{s}
}
\description{
Supporting function for gradient computation in \code{\link{SSDL}[CDL]}.
}
\keyword{internal}
