% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradient}
\alias{gradient}
\title{Supporting function for the gradient computation.}
\usage{
gradient(S, y, W)
}
\arguments{
\item{S}{is a matrix}

\item{y}{is a vector}

\item{W}{is a frequency matrix}
}
\value{
a gradient stored in the matrix \code{S}.
}
\description{
The gradient computation using a parallel application of the function \code{\link{G_fun_cpp}}
to a vector \code{y}, which is defined as a difference between a data sketch and
the decomposition sketch, \emph{i.e.} \eqn{SK - SK(D\cdot A)}.
}
\seealso{
\code{\link{Gradient_D_cpp_parallel}}
}
\keyword{internal}
