% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gradient_D_cpp}
\alias{Gradient_D_cpp}
\title{Gradient D}
\usage{
Gradient_D_cpp(D, A, W, SK)
}
\arguments{
\item{D}{is a dictionary \eqn{s \times K}.}

\item{A}{is a code matrix \eqn{K \times n}.}

\item{W}{is a frequency matrix \eqn{m \times s} with frequency vectors in matrix rows.}

\item{SK}{is a data sketch. It is a \eqn{2m}-dimensional vector.}
}
\value{
a gradient matrix
}
\description{
Gradient with respect to dictionary using sparse product
}
\keyword{internal}
