% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TV_initialization.R
\name{make_copy}
\alias{make_copy}
\title{function makes copy of the selected signal}
\usage{
make_copy(x, n_copies = 2)
}
\arguments{
\item{x}{is a selected signal pattern}

\item{n_copies}{is a number of the copies.}
}
\value{
a matrix, which contains \code{n_copies} of the patters signal in the columns.
}
\description{
function makes copy of the selected signal
}
\keyword{internal}
