% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rowsums_cpp_parallel}
\alias{Rowsums_cpp_parallel}
\title{Rowsums_cpp_parallel}
\usage{
Rowsums_cpp_parallel(P, SK)
}
\arguments{
\item{P}{is a matrix}

\item{SK}{is a data sketch}
}
\value{
a vector of matrix row sums minus a data sketch
}
\description{
Parallel implementation of a matrix rows' summation and a sketch vector subtraction
}
\seealso{
\code{\link[SSDL]{Rowsums_cpp}}
}
\keyword{internal}
