% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sparse_prod}
\alias{Sparse_prod}
\title{Sparse matrix product}
\usage{
Sparse_prod(D, Z)
}
\arguments{
\item{D}{is a matrix}

\item{Z}{is a sparse matrix}
}
\value{
a product matrix
}
\description{
Implementation of the product a matrix by a sparse matrix
}
\keyword{internal}
