% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sparse_prod_parallel}
\alias{Sparse_prod_parallel}
\title{Sparse matrix product}
\usage{
Sparse_prod_parallel(D, Z)
}
\arguments{
\item{D}{is a matrix}

\item{Z}{is a matrix}
}
\value{
a product matrix
}
\description{
Parallel computation of a product of a matrix by a sparse matrix
}
\keyword{internal}
