% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sparse_prod_row}
\alias{Sparse_prod_row}
\title{Sparse product by rows}
\usage{
Sparse_prod_row(D, Z, CosSin = FALSE)
}
\arguments{
\item{D}{is a matrix}

\item{Z}{is a matrix}

\item{CosSin}{indicates whether to compute cos and sin of the resulting matrix}
}
\value{
a matrix
}
\description{
Implementation sparse matrix product by rows
}
\keyword{internal}
