% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TV_initialization.R
\name{step}
\alias{step}
\title{Additional function for the TV norm initialization procedure.}
\usage{
step(x, n_copies, step = 1)
}
\arguments{
\item{x}{is a selected signal pattern}

\item{n_copies}{is a number of the copies.}

\item{step}{is a step size. The default value is 1.}
}
\value{
three dimensional vector with the optimal step size, lower and upper step size bounds.
}
\description{
Additional function for the TV norm initialization procedure.
}
\keyword{internal}
